% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_by_Method.R
\name{CNV.by.method}
\alias{CNV.by.method}
\title{CNV generated by given method}
\usage{
CNV.by.method(CNV.input, gene.name, pids, title, legend, legend.names,
  out.dir, file.type, pixel.per.cnv, color, display, cnv.type, gene.anno,
  start.gene, end.gene, sort.method, color.method, zoomed, path, format,
  SaveAsObject, orient)
}
\arguments{
\item{CNV.input}{Grange object}

\item{gene.name}{string}

\item{legend}{int}

\item{legend.names}{string}

\item{out.dir}{string}

\item{file.type}{string}

\item{pixel.per.cnv}{int}

\item{color}{list}

\item{display}{string}

\item{gene.anno}{string}

\item{start.gene}{string}

\item{end.gene}{string}

\item{sort.method}{string}

\item{color.method}{string}

\item{pid}{string}
}
\value{
list variable
}
\description{
CNV generated by given method
}
