% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TornadoPlots_CNV_single.R
\docType{methods}
\name{TornadoPlots,CNV_single-method}
\alias{TornadoPlots,CNV_single-method}
\title{Plot Tornado Plots}
\usage{
\S4method{TornadoPlots}{CNV_single}(object, gene.name, pids, title, legend,
  legend.names, out.dir, file.type, pixel.per.cnv, color, display,
  gene.anno, start.gene, end.gene, color.method, sort.method, SaveAsObject,
  multi_panel, zoomed, font.size.factor, path, format, orient)
}
\arguments{
\item{gene.name}{gene name. Character.}

\item{pids}{the given patient id. Character.}

\item{legend}{type of legend display. Character.}

\item{legend.names}{string}

\item{color}{colors of tornado.List.}

\item{color.method}{the method for coloring the CNVs. Character.}

\item{sort.method}{the method for coloring the CNVs. Character.}

\item{SaveAsObject}{if true, the plot will be saved in R object. If false, it will be saved in a file. Boolean.}

\item{multi_panel}{If true, a multiple panel plot will be shown.}

\item{zoomed}{if true, the plot will be shown in zoomed-in sight. Boolean.}

\item{font.size.factor}{multiply for rescale font size. Double.}

\item{path}{path of output plot. character.}

\item{orient}{vertical or horizontal arrange CNVs and chromosome. It should be "v" or "h".}

\item{CNV.input}{Output of MakeData function, Grange object}

\item{formath}{output plot file type. character.}
}
\description{
Plot Tornado Plots
}
