% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOrnadoPlots_CNV_twin.R
\docType{methods}
\name{TornadoPlots,CNV_twin-method}
\alias{TornadoPlots,CNV_twin-method}
\title{Plot Tornado Twin Plots}
\usage{
\S4method{TornadoPlots}{CNV_twin}(object, gene.name, title, legend.type,
  legend.names, pixel.per.cnv, color, cnv.type_1, cnv.type_2, gene.anno,
  start.gene, end.gene, color.method, sort.method, SaveAsObject,
  multi_panel, font.size.factor, file, format, path, orient, zoomed,
  drop.low.amp)
}
\arguments{
\item{gene.name}{gene name. Character.}

\item{legend.names}{type of legend display. Character.}

\item{color}{colors of tornado.List.}

\item{color.method}{the method for coloring the CNVs. Character.}

\item{sort.method}{the method for coloring the CNVs. Character.}

\item{SaveAsObject}{if true, the plot will be saved in R object. If false, it will be saved in a file. Boolean.}

\item{multi_panel}{If true, a multiple panel plot will be shown.}

\item{font.size.factor}{multiply for rescale font size. Double.}

\item{path}{path of output plot. character.}

\item{orient}{vertical or horizontal arrange CNVs and chromosome. It should be "v" or "h".}

\item{zoomed}{if true, the plot will be shown in zoomed-in sight. Boolean.}

\item{drop.low.amp}{If it is TRUE (default), the amplifications with CNV<5 are not shown in plot. Boolean.}

\item{CNV.input}{Grange object}

\item{legend}{type of legend display. Character.}

\item{formath}{output plot file type. character.}
}
\description{
Plot Tornado Twin Plots
}
