% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_by_Method.R
\name{CNV.by.method}
\alias{CNV.by.method}
\title{CNV generated by given method}
\usage{
CNV.by.method(CNV.input, gene.name, title, legend, out.dir, file.type,
  pixel.per.cnv, color, cnv.type, gene.anno, start.gene, end.gene,
  sort.method, color.method, zoomed, path, format, SaveAsObject, orient,
  drop.low.amp)
}
\arguments{
\item{CNV.input}{Grange object}

\item{gene.name}{string}

\item{legend}{int}

\item{out.dir}{string}

\item{file.type}{string}

\item{pixel.per.cnv}{int}

\item{color}{list}

\item{gene.anno}{string}

\item{start.gene}{string}

\item{end.gene}{string}

\item{sort.method}{string}

\item{color.method}{string}

\item{pid}{string}
}
\value{
list variable
}
\description{
CNV generated by given method
}
