% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeData.R
\name{MakeData}
\alias{MakeData}
\title{Generate input data for tornado plot}
\usage{
MakeData(CNV, gene_name_1, gene_name_2, score.type, max.length,
  score.method, cohort_thredshold)
}
\arguments{
\item{CNV}{data frame with six columns,Chromosome,Start,End,Score,Gene,Cohort,PID}

\item{gene_name_1}{character, the name of first gene}

\item{gene_name_2}{character, the name of second gene}

\item{CNV}{data frame with six columns,Chromosome,Start,End,Score,Gene,Cohort,PID}

\item{max.length}{numeric, neglect the events above the max length.}

\item{score.method}{character, the method of focallity score, could be "edge" or "normal". default "edge".}

\item{cohort_thredshold}{numeric,the minimum size of cohort. default 0.}



}
\description{
Generate input data for tornado plot
}
