% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{paintCytobands}
\alias{lengthChromosome}
\alias{plot}
\alias{legend}
\alias{axis}
\alias{boxplot}
\alias{lines}
\alias{par}
\alias{pie}
\alias{polygon}
\alias{text}
\alias{plot.new}
\alias{rect}
\alias{colorRampPalette}
\alias{dev.off}
\alias{gray.colors}
\alias{palette}
\alias{pdf}
\alias{rainbow}
\alias{makeGRangesFromDataFrame}
\alias{GRanges}
\alias{seqnames}
\alias{start}
\alias{end}
\alias{grid.arrange}
\alias{rasterGrob}
\alias{readTIFF}
\alias{IRanges}
\alias{subsetByOverlaps}
\alias{Rle}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{GenomicRanges}{\code{\link[GenomicRanges]{makeGRangesFromDataFrame}}, \code{\link[GenomicRanges]{GRanges}}, \code{\link[GenomicRanges]{seqnames}}, \code{\link[GenomicRanges]{start}}, \code{\link[GenomicRanges]{end}}}

  \item{graphics}{\code{\link[graphics]{plot}}, \code{\link[graphics]{legend}}, \code{\link[graphics]{axis}}, \code{\link[graphics]{boxplot}}, \code{\link[graphics]{lines}}, \code{\link[graphics]{par}}, \code{\link[graphics]{pie}}, \code{\link[graphics]{polygon}}, \code{\link[graphics]{text}}, \code{\link[graphics]{plot.new}}, \code{\link[graphics]{rect}}}

  \item{grDevices}{\code{\link[grDevices]{colorRampPalette}}, \code{\link[grDevices]{dev.off}}, \code{\link[grDevices]{gray.colors}}, \code{\link[grDevices]{palette}}, \code{\link[grDevices]{pdf}}, \code{\link[grDevices]{rainbow}}}

  \item{grid}{\code{\link[grid]{rasterGrob}}}

  \item{gridExtra}{\code{\link[gridExtra]{grid.arrange}}}

  \item{IRanges}{\code{\link[IRanges]{IRanges}}, \code{\link[IRanges]{subsetByOverlaps}}}

  \item{quantsmooth}{\code{\link[quantsmooth]{paintCytobands}}, \code{\link[quantsmooth]{lengthChromosome}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{Rle}}}

  \item{tiff}{\code{\link[tiff]{readTIFF}}}
}}

