% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TornadoPlots_CNV_single.R
\docType{methods}
\name{TornadoPlots,CNV_single-method}
\alias{TornadoPlots,CNV_single-method}
\title{Plot Tornado Plots}
\usage{
\S4method{TornadoPlots}{CNV_single}(object, gene.name, title, legend,
  pixel.per.cnv, color, gene.anno, start.gene, end.gene, color.method,
  sort.method, SaveAsObject, drop.low.amp, multi_panel, zoomed,
  font.size.factor, path, format, orient)
}
\arguments{
\item{gene.name}{gene name. Character.}

\item{title}{the title shown in the plots. Character.}

\item{legend}{type of legend display. Character.}

\item{pixel.per.cnv}{thickness of CNV lines in pixel. Int.}

\item{color}{colors of tornado.List.}

\item{color.method}{the method for coloring the CNVs. Character.}

\item{sort.method}{the method for coloring the CNVs. Character.}

\item{SaveAsObject}{if true, the plot will be saved in R object. If false, it will be saved in a file. Boolean.}

\item{drop.low.amp}{If it is TRUE (default), the amplifications with CNV<5 are not shown in plot. Boolean.}

\item{multi_panel}{If true, a multiple panel plot will be shown.}

\item{zoomed}{indicates if plot will be shown in zoomed-in sight. It should be "global","region" or "gene".Character}

\item{font.size.factor}{multiply for rescale font size. Double.}

\item{path}{path of output plot. character.}

\item{orient}{vertical or horizontal arrange CNVs and chromosome. It should be "v" or "h".}

\item{CNV.input}{Output of MakeData function, Grange object}

\item{formath}{output plot file type. character.}
}
\description{
Plot Tornado Plots
}
