% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TornadoPlots_CNV_single.R
\docType{methods}
\name{TornadoPlots,CNV_single-method}
\alias{TornadoPlots,CNV_single-method}
\title{Plot Tornado Plots}
\usage{
\S4method{TornadoPlots}{CNV_single}(object, gene.name, pids, title, legend,
  legend.names, out.dir, file.type, pixel.per.cnv, color, display,
  gene.anno, start.gene, end.gene, color.method, sort.method, SaveAsObject,
  multi_panel, zoomed, font.size.factor)
}
\arguments{
\item{gene.name}{gene name, string}

\item{pids}{the given patient id, string}

\item{legend}{type of legend display, int}

\item{legend.names}{string}

\item{out.dir}{path of output plot, string}

\item{file.type}{output plot file type, string}

\item{pixel.per.cnv}{int}

\item{color}{colors of tornado, list}

\item{display}{string}

\item{gene.anno}{string}

\item{start.gene}{string}

\item{end.gene}{string}

\item{color.method}{string}

\item{sort.method}{string}

\item{CNV.input}{Output of MakeData function, Grange object}
}
\description{
Plot Tornado Plots
}
