% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TornadoPlots_CNV_twin.R
\docType{methods}
\name{TornadoPlots,CNV_twin-method}
\alias{TornadoPlots,CNV_twin-method}
\title{Plot Tornado Twin Plots}
\usage{
\S4method{TornadoPlots}{CNV_twin}(object, gene.name, pids, title,
  legend.type, legend.names, out.dir, file.type, pixel.per.cnv, color,
  display, cnv.type_1, cnv.type_2, gene.anno, start.gene, end.gene,
  color.method, sort.method, SaveAsObject, multi_panel)
}
\arguments{
\item{gene.name}{string}

\item{legend.names}{string}

\item{out.dir}{string}

\item{file.type}{string}

\item{pixel.per.cnv}{int}

\item{color}{list}

\item{display}{string}

\item{gene.anno}{string}

\item{start.gene}{string}

\item{end.gene}{string}

\item{color.method}{string}

\item{sort.method}{string}

\item{CNV.input}{Grange object}

\item{pid}{string}

\item{legend}{int}
}
\description{
Plot Tornado Twin Plots
}
